Move some functions depending on runtime_shape to runtime_shape.h
